fix,update(mkfile): updated mkfile build system, using cccmk

+++ mkfile/packages/SDL2.mk
@@ -34,49 +34,52 @@ PACKAGE_SDL2_PKG_INSTALL = \
 	@$(call print_warning,"Unsupported platform: SDL2 must be configured manually")
 
 ifeq ($(OSMODE),other)
-else ifneq ($(filter $(OSMODE), win32 win64),)
+$(error Unsupported platform: external package 'SDL2' must be configured manually)
+endif
+ifeq ($(OSMODE),windows)
 	PACKAGE_SDL2_PKG = SDL2-devel-$(PACKAGE_SDL2_VERSION)-mingw.tar.gz
+	PACKAGE_SDL2_GETVERSIONS = grep 'mingw.tar.gz'
 	PACKAGE_SDL2_PKG_INSTALL = \
 		tar -xf $(PACKAGE_SDL2_PKG) --directory=$(PACKAGE_SDL2_DIR) ; \
 		mv -f $(PACKAGE_SDL2_DIR)SDL2-$(PACKAGE_SDL2_VERSION)/* $(PACKAGE_SDL2_DIR) && \
 		rmdir $(PACKAGE_SDL2_DIR)SDL2-$(PACKAGE_SDL2_VERSION) ; \
-		{ mkdir -p $(PACKAGE_SDL2_DIR)bin/win32/ && mv $(PACKAGE_SDL2_DIR)i686-w64-mingw32/*   $(PACKAGE_SDL2_DIR)bin/win32/ && rmdir $(PACKAGE_SDL2_DIR)i686-w64-mingw32   ; } ; \
-		{ mkdir -p $(PACKAGE_SDL2_DIR)bin/win64/ && mv $(PACKAGE_SDL2_DIR)x86_64-w64-mingw32/* $(PACKAGE_SDL2_DIR)bin/win64/ && rmdir $(PACKAGE_SDL2_DIR)x86_64-w64-mingw32 ; } ;
-	PACKAGE_SDL2_GETVERSIONS = \
-		grep 'mingw.tar.gz'
-else ifeq ($(OSMODE),macos)
+		{ mv $(PACKAGE_SDL2_DIR)i686-w64-mingw32/   $(PACKAGE_SDL2_DIR)bin/windows_x86-32/ ; } ; \
+		{ mv $(PACKAGE_SDL2_DIR)x86_64-w64-mingw32/ $(PACKAGE_SDL2_DIR)bin/windows_x86-64/ ; } ; \
+
+endif
+ifeq ($(OSMODE),macos)
 	PACKAGE_SDL2_PKG = SDL2-$(PACKAGE_SDL2_VERSION).dmg
+	PACKAGE_SDL2_GETVERSIONS = grep '.dmg'
 	PACKAGE_SDL2_PKG_INSTALL = \
 		listing=`hdiutil attach $(PACKAGE_SDL2_PKG) | grep Volumes` ; \
 		volume=`echo "$$listing" | cut -f 3` ; \
 		cp -rf "$$volume"/SDL2.framework $(PACKAGE_SDL2_BIN)$(PACKAGE_SDL2_LIBMODE) ; \
-		hdiutil detach `echo "$$listing" | cut -f 1`
-	PACKAGE_SDL2_GETVERSIONS = \
-		grep '.dmg'
-else ifeq ($(OSMODE),linux)
+		hdiutil detach `echo "$$listing" | cut -f 1` ; \
+
+endif
+ifeq ($(OSMODE),linux)
 	PACKAGE_SDL2_PKG = SDL2-$(PACKAGE_SDL2_VERSION).zip
+	PACKAGE_SDL2_GETVERSIONS = grep '.zip'
 	PACKAGE_SDL2_PKG_INSTALL = \
 		unzip $(PACKAGE_SDL2_PKG) -d $(PACKAGE_SDL2_DIR) ; \
 		mv -f $(PACKAGE_SDL2_DIR)SDL2-$(PACKAGE_SDL2_VERSION)/* $(PACKAGE_SDL2_BIN) && \
 		rmdir $(PACKAGE_SDL2_DIR)SDL2-$(PACKAGE_SDL2_VERSION) ; \
 		cd $(PACKAGE_SDL2_BIN) ; \
-		./configure && make
-	PACKAGE_SDL2_GETVERSIONS = \
-		grep '.zip'
-else
-$(error Unsupported platform: external package 'SDL2' must be configured manually)
+		./configure && make ; \
+
 endif
 
 
 
 #! The shell command to retrieve and output list of newer versions, if any
+#! @param $(1)	The "current version number" to compare to
 package_SDL2_checkupdates = \
 	curl --silent $(PACKAGE_SDL2_URL) \
 	| grep 'SDL2' \
 	| cut -d'"' -f 8 \
 	| $(PACKAGE_SDL2_GETVERSIONS) \
 	| sort --version-sort \
-	| awk -v found=0 '\
+	| awk -v found=$(1) '\
 	{\
 		if (/$(PACKAGE_SDL2_VERSION)/) { found = 1 }\
 		else if (found) { found += 1; print; }\
@@ -88,6 +91,10 @@ package_SDL2_checkupdates = \
 .PHONY:\
 package-SDL2 #! downloads the package, according to the version number set
 package-SDL2:
+ifeq ($(strip $(PACKAGE_SDL2_VERSION)),?)
+	$(call print_message,"No specific version set - getting latest version...")
+	$(eval PACKAGE_SDL2_VERSION := $(shell $(call package_SDL2_checkupdates,1) | tail -1))
+endif
 	@$(call packages_setversion,$(PACKAGE_SDL2),$(PACKAGE_SDL2_VERSION))
 	@$(call print_message,"Downloading package: $(PACKAGE_SDL2)@$(PACKAGE_SDL2_VERSION)...")
 	@curl $(PACKAGE_SDL2_URL)$(PACKAGE_SDL2_PKG) --progress-bar --output $(PACKAGE_SDL2_PKG)
@@ -102,9 +109,12 @@ package-SDL2:
 update-SDL2 #! updates the package to the latest version
 update-SDL2:
 	@$(call print_message,"Checking new versions for package: $(PACKAGE_SDL2)...")
+	@if [ "$(PACKAGE_SDL2_VERSION)" = "?" ] || ! [ -d "$(PACKAGE_SDL2_DIR)" ] ; then \
+		$(MAKE) package-SDL2 ; \
+	fi
 	@echo "=> Current version is: $(PACKAGE_SDL2_VERSION)"
-	@new_versions=`$(call package_SDL2_checkupdates)` ; \
-	if [ -z "$${new_versions}" ]; then \
+	@new_versions=`$(call package_SDL2_checkupdates,0)` ; \
+	if [ -z "$${new_versions}" ] ; then \
 		printf "Newest version already set.\n" ; \
 	else \
 		newer_version=`echo "$${new_versions}" | tail -1` ; \
